Move maze helper function to their own file.
